feat(app): move frames to a file for each one
